Move and rename kernel profiler APIs
