+++ python/taichi/__init__.py
@@ -22,7 +22,16 @@ __deprecated_names__ = {
     'SOA': 'Layout.SOA',
     'AOS': 'Layout.AOS',
     'print_profile_info': 'profiler.print_scoped_profiler_info',
-    'clear_profile_info': 'profiler.clear_scoped_profiler_info'
+    'clear_profile_info': 'profiler.clear_scoped_profiler_info',
+    'CuptiMetric': 'profiler.CuptiMetric',
+    'get_predefined_cupti_metrics': 'profiler.get_predefined_cupti_metrics',
+    'print_kernel_profile_info': 'profiler.print_kernel_profiler_info',
+    'query_kernel_profile_info': 'profiler.query_kernel_profiler_info',
+    'clear_kernel_profile_info': 'profiler.clear_kernel_profiler_info',
+    'kernel_profiler_total_time': 'profiler.get_kernel_profiler_total_time',
+    'set_kernel_profiler_toolkit': 'profiler.set_kernel_profiler_toolkit',
+    'set_kernel_profile_metrics': 'profiler.set_kernel_profiler_metrics',
+    'collect_kernel_profile_metrics': 'profiler.collect_kernel_profiler_metrics'
 }
 
 if sys.version_info.minor < 7:
