+++ python/taichi/lang/__init__.py
@@ -24,9 +24,6 @@ from taichi.lang.sort import parallel_sort
 from taichi.lang.source_builder import SourceBuilder
 from taichi.lang.struct import Struct, StructField
 from taichi.lang.type_factory_impl import type_factory
-from taichi.profiler import KernelProfiler, get_default_kernel_profiler
-from taichi.profiler.kernel_metrics import (CuptiMetric, default_cupti_metrics,
-                                            get_predefined_cupti_metrics)
 from taichi.tools.util import set_gdb_trigger, warning
 from taichi.types.annotations import any_arr, ext_arr, template
 from taichi.types.primitive_types import f16, f32, f64, i32, i64, u32, u64
