+++ python/taichi/profiler/kernel_profiler.py
@@ -40,7 +40,7 @@ class KernelProfiler:
     """Kernel profiler of Taichi.
 
     Kernel profiler acquires kernel profiling records from backend, counts records in Python scope,
-    and prints the results to the console by :func:`~taichi.profiler.kernelprofiler.KernelProfiler.print_info`.
+    and prints the results to the console by :func:`~taichi.profiler.kernel_profiler.KernelProfiler.print_info`.
 
     ``KernelProfiler`` now support detailed low-level performance metrics (such as memory bandwidth consumption) in its advanced mode.
     This mode is only available for the CUDA backend with CUPTI toolkit, i.e. you need ``ti.init(kernel_profiler=True, arch=ti.cuda)``.
@@ -59,7 +59,7 @@ class KernelProfiler:
     # public methods
 
     def set_kernel_profiler_mode(self, mode=False):
-        """Turn on or off :class:`~taichi.profiler.kernelprofiler.KernelProfiler`."""
+        """Turn on or off :class:`~taichi.profiler.kernel_profiler.KernelProfiler`."""
         if type(mode) is bool:
             self._profiling_mode = mode
         else:
@@ -68,7 +68,7 @@ class KernelProfiler:
             )
 
     def get_kernel_profiler_mode(self):
-        """Get status of :class:`~taichi.profiler.kernelprofiler.KernelProfiler`."""
+        """Get status of :class:`~taichi.profiler.kernel_profiler.KernelProfiler`."""
         return self._profiling_mode
 
     def set_toolkit(self, toolkit_name='default'):
@@ -97,7 +97,7 @@ class KernelProfiler:
         return self._total_time_ms / 1000  # ms to s
 
     def clear_info(self):
-        """Clear all records both in front-end :class:`~taichi.profiler.kernelprofiler.KernelProfiler` and back-end instance ``KernelProfilerBase``.
+        """Clear all records both in front-end :class:`~taichi.profiler.kernel_profiler.KernelProfiler` and back-end instance ``KernelProfilerBase``.
 
         Note:
             The values of ``self._profiling_mode`` and ``self._metric_list`` will not be cleared.
@@ -113,7 +113,7 @@ class KernelProfiler:
         return None
 
     def query_info(self, name):
-        """For docsting of this function, see :func:`~taichi.lang.query_kernel_profile_info`."""
+        """For docstring of this function, see :func:`~taichi.profiler.query_kernel_profiler_info`."""
         if self._check_not_turned_on_with_warning_message():
             return None
         self._update_records()  # kernel records
@@ -122,7 +122,7 @@ class KernelProfiler:
         return impl.get_runtime().prog.query_kernel_profile_info(name)
 
     def set_metrics(self, metric_list=default_cupti_metrics):
-        """For docsting of this function, see :func:`~taichi.lang.set_kernel_profile_metrics`."""
+        """For docstring of this function, see :func:`~taichi.profiler.set_kernel_profiler_metrics`."""
         if self._check_not_turned_on_with_warning_message():
             return None
         self._metric_list = metric_list
@@ -137,7 +137,7 @@ class KernelProfiler:
     def collect_metrics_in_context(self, metric_list=default_cupti_metrics):
         """This function is not exposed to user now.
 
-        For usage of this function, see :func:`~taichi.lang.collect_kernel_profile_metrics`.
+        For usage of this function, see :func:`~taichi.profiler.collect_kernel_profiler_metrics`.
         """
         if self._check_not_turned_on_with_warning_message():
             return None
@@ -154,7 +154,7 @@ class KernelProfiler:
     def print_info(self, mode=COUNT):
         """Print the profiling results of Taichi kernels.
 
-        For usage of this function, see :func:`~taichi.lang.print_kernel_profile_info`.
+        For usage of this function, see :func:`~taichi.profiler.print_kernel_profiler_info`.
 
         Args:
             mode (str): the way to print profiling results.
@@ -187,7 +187,7 @@ class KernelProfiler:
         return False
 
     def _clear_frontend(self):
-        """Clear member variables in :class:`~taichi.profiler.kernelprofiler.KernelProfiler`.
+        """Clear member variables in :class:`~taichi.profiler.kernel_profiler.KernelProfiler`.
 
         Note:
             The values of ``self._profiling_mode`` and ``self._metric_list`` will not be cleared.
@@ -353,3 +353,235 @@ def get_default_kernel_profiler():
     For data retention purposes, multiple instances will be considered in the future.
     """
     return _ti_kernel_profiler
+
+
+def print_kernel_profiler_info(mode='count'):
+    """Print the profiling results of Taichi kernels.
+
+    To enable this profiler, set ``kernel_profiler=True`` in ``ti.init()``.
+    ``'count'`` mode: print the statistics (min,max,avg time) of launched kernels,
+    ``'trace'`` mode: print the records of launched kernels with specific profiling metrics (time, memory load/store and core utilization etc.),
+    and defaults to ``'count'``.
+
+    Args:
+        mode (str): the way to print profiling results.
+
+    Example::
+
+        >>> import taichi as ti
+
+        >>> ti.init(ti.cpu, kernel_profiler=True)
+        >>> var = ti.field(ti.f32, shape=1)
+
+        >>> @ti.kernel
+        >>> def compute():
+        >>>     var[0] = 1.0
+
+        >>> compute()
+        >>> ti.profiler.print_kernel_profiler_info()
+        >>> # equivalent calls :
+        >>> # ti.profiler.print_kernel_profiler_info('count')
+
+        >>> ti.profiler.print_kernel_profiler_info('trace')
+
+    Note:
+        Currently the result of `KernelProfiler` could be incorrect on OpenGL
+        backend due to its lack of support for `ti.sync()`.
+
+        For advanced mode of `KernelProfiler`, please visit https://docs.taichi.graphics/docs/lang/articles/misc/profiler#advanced-mode.
+    """
+    get_default_kernel_profiler().print_info(mode)
+
+
+def query_kernel_profiler_info(name):
+    """Query kernel elapsed time(min,avg,max) on devices using the kernel name.
+
+    To enable this profiler, set `kernel_profiler=True` in `ti.init`.
+
+    Args:
+        name (str): kernel name.
+
+    Returns:
+        KernelProfilerQueryResult (class): with member variables(counter, min, max, avg)
+
+    Example::
+
+        >>> import taichi as ti
+
+        >>> ti.init(ti.cpu, kernel_profiler=True)
+        >>> n = 1024*1024
+        >>> var = ti.field(ti.f32, shape=n)
+
+        >>> @ti.kernel
+        >>> def fill():
+        >>>     for i in range(n):
+        >>>         var[i] = 0.1
+
+        >>> fill()
+        >>> ti.profiler.clear_kernel_profiler_info() #[1]
+        >>> for i in range(100):
+        >>>     fill()
+        >>> query_result = ti.profiler.query_kernel_profiler_info(fill.__name__) #[2]
+        >>> print("kernel excuted times =",query_result.counter)
+        >>> print("kernel elapsed time(min_in_ms) =",query_result.min)
+        >>> print("kernel elapsed time(max_in_ms) =",query_result.max)
+        >>> print("kernel elapsed time(avg_in_ms) =",query_result.avg)
+
+    Note:
+        [1] To get the correct result, query_kernel_profiler_info() must be used in conjunction with
+        clear_kernel_profiler_info().
+
+        [2] Currently the result of `KernelProfiler` could be incorrect on OpenGL
+        backend due to its lack of support for `ti.sync()`.
+    """
+    return get_default_kernel_profiler().query_info(name)
+
+
+def clear_kernel_profiler_info():
+    """Clear all KernelProfiler records."""
+    get_default_kernel_profiler().clear_info()
+
+
+def get_kernel_profiler_total_time():
+    """Get elapsed time of all kernels recorded in KernelProfiler.
+
+    Returns:
+        time (float): total time in second.
+    """
+    return get_default_kernel_profiler().get_total_time()
+
+
+def set_kernel_profiler_toolkit(toolkit_name='default'):
+    """Set the toolkit used by KernelProfiler.
+
+    Currently, we only support toolkits: ``'default'`` and ``'cupti'``.
+
+    Args:
+        toolkit_name (str): string of toolkit name.
+
+    Returns:
+        status (bool): whether the setting is successful or not.
+
+    Example::
+
+        >>> import taichi as ti
+
+        >>> ti.init(arch=ti.cuda, kernel_profiler=True)
+        >>> x = ti.field(ti.f32, shape=1024*1024)
+
+        >>> @ti.kernel
+        >>> def fill():
+        >>>     for i in x:
+        >>>         x[i] = i
+
+        >>> ti.profiler.set_kernel_profiler_toolkit('cupti')
+        >>> for i in range(100):
+        >>>     fill()
+        >>> ti.profiler.print_kernel_profiler_info()
+
+        >>> ti.profiler.set_kernel_profiler_toolkit('default')
+        >>> for i in range(100):
+        >>>     fill()
+        >>> ti.profiler.print_kernel_profiler_info()
+    """
+    return get_default_kernel_profiler().set_toolkit(toolkit_name)
+
+
+def set_kernel_profiler_metrics(metric_list=default_cupti_metrics):
+    """Set metrics that will be collected by the CUPTI toolkit.
+
+    Args:
+        metric_list (list): a list of :class:`~taichi.profiler.CuptiMetric()` instances, default value: :data:`~taichi.profiler.kernel_metrics.default_cupti_metrics`.
+
+    Example::
+
+        >>> import taichi as ti
+
+        >>> ti.init(kernel_profiler=True, arch=ti.cuda)
+        >>> ti.profiler.set_kernel_profiler_toolkit('cupti')
+        >>> num_elements = 128*1024*1024
+
+        >>> x = ti.field(ti.f32, shape=num_elements)
+        >>> y = ti.field(ti.f32, shape=())
+        >>> y[None] = 0
+
+        >>> @ti.kernel
+        >>> def reduction():
+        >>>     for i in x:
+        >>>         y[None] += x[i]
+
+        >>> # In the case of not pramater, Taichi will print its pre-defined metrics list
+        >>> ti.profiler.get_predefined_cupti_metrics()
+        >>> # get Taichi pre-defined metrics
+        >>> profiling_metrics = ti.profiler.get_predefined_cupti_metrics('shared_access')
+
+        >>> global_op_atom = ti.profiler.CuptiMetric(
+        >>>     name='l1tex__t_set_accesses_pipe_lsu_mem_global_op_atom.sum',
+        >>>     header=' global.atom ',
+        >>>     format='    {:8.0f} ')
+        >>> # add user defined metrics
+        >>> profiling_metrics += [global_op_atom]
+
+        >>> # metrics setting will be retained until the next configuration
+        >>> ti.profiler.set_kernel_profile_metrics(profiling_metrics)
+        >>> for i in range(16):
+        >>>     reduction()
+        >>> ti.profiler.print_kernel_profiler_info('trace')
+
+    Note:
+        Metrics setting will be retained until the next configuration.
+    """
+    get_default_kernel_profiler().set_metrics(metric_list)
+
+
+@contextmanager
+def collect_kernel_profiler_metrics(metric_list=default_cupti_metrics):
+    """Set temporary metrics that will be collected by the CUPTI toolkit within this context.
+
+    Args:
+        metric_list (list): a list of :class:`~taichi.profiler.CuptiMetric()` instances, default value: :data:`~taichi.profiler.kernel_metrics.default_cupti_metrics`.
+
+    Example::
+
+        >>> import taichi as ti
+
+        >>> ti.init(kernel_profiler=True, arch=ti.cuda)
+        >>> ti.profiler.set_kernel_profiler_toolkit('cupti')
+        >>> num_elements = 128*1024*1024
+
+        >>> x = ti.field(ti.f32, shape=num_elements)
+        >>> y = ti.field(ti.f32, shape=())
+        >>> y[None] = 0
+
+        >>> @ti.kernel
+        >>> def reduction():
+        >>>     for i in x:
+        >>>         y[None] += x[i]
+
+        >>> # In the case of not pramater, Taichi will print its pre-defined metrics list
+        >>> ti.profiler.get_predefined_cupti_metrics()
+        >>> # get Taichi pre-defined metrics
+        >>> profiling_metrics = ti.profiler.get_predefined_cupti_metrics('device_utilization')
+
+        >>> global_op_atom = ti.profiler.CuptiMetric(
+        >>>     name='l1tex__t_set_accesses_pipe_lsu_mem_global_op_atom.sum',
+        >>>     header=' global.atom ',
+        >>>     format='    {:8.0f} ')
+        >>> # add user defined metrics
+        >>> profiling_metrics += [global_op_atom]
+
+        >>> # metrics setting is temporary, and will be clear when exit from this context.
+        >>> with ti.profiler.collect_kernel_profiler_metrics(profiling_metrics):
+        >>>     for i in range(16):
+        >>>         reduction()
+        >>>     ti.profiler.print_kernel_profiler_info('trace')
+
+    Note:
+        The configuration of the ``metric_list`` will be clear when exit from this context.
+    """
+    get_default_kernel_profiler().set_metrics(metric_list)
+    yield get_default_kernel_profiler()
+    get_default_kernel_profiler().set_metrics()
+
+
+__all__ = ['clear_kernel_profiler_info', 'collect_kernel_profiler_metrics', 'get_kernel_profiler_total_time', 'print_kernel_profiler_info', 'query_kernel_profiler_info', 'set_kernel_profiler_metrics', 'set_kernel_profiler_toolkit']
